deploy app to gds-way and redirect traffic from old app
This deploys the site to a new app called `gds-way`.  It also sets up
a static nginx at the old app to redirect all traffic to the new URLs.

I tested this locally by deploying the redirect to a site called
`gds-tech-docs-test` and verifying that the redirects worked as
expected.