Add some basic guidance on naming things
Rather than including guidance on naming things everywhere those names need to be considered (e.g, naming GitHub repos and in naming packages), add some general documentation on how we name things at GDS which we can refer to as appropriate.

This is loosely based on an existing [GOV.UK RFC] as well as my personal understanding of how we name things.

[GOV.UK RFC]: https://github.com/alphagov/govuk-rfcs/blob/master/rfc-063-naming-new-apps-gems-on-gov-uk.md