Move logstash filters into PaaS Tech Docs
Until now the GDS Way has included very basic information on how to
configure logit for use with Cloud Foundry. We've released our own
documentation on this, using the same logstash filters and with more
explanation.

This pull request removes the now-duplicate information from the
GDS Way and instead links to the PaaS Documentation.