Add .editorconfig file with guidelines from Java style guide
The indent_style, indent_size and continuation_indent_size properties match the
IntelliJ defaults, per the style guide’s “Use the default code indenting in
IntelliJ”. The insert_final_newline property matches the style guide’s
instruction to “Ensure line feeds are present at the ends of files”. The
end_of_line and charset properties don’t match anything in the style guide but
I put them in because it’s what we do anyway.

After deliberately changing the IntelliJ settings so they didn’t match, I
dropped the .editorconfig file in the root of one of our Java projects. It
seemed to work. I also installed the plugins for Eclipse and Sublime Text and
tried the .editorconfig file with those: both worked (with some quirks) but
they don’t support all of the properties.

Looking through the list of other EditorConfig properties, we might want to
specify something for trim_trailing_whitespace, max_line_length,
curly_bracket_next_line, spaces_around_operators and indent_brace_style.
However, IntelliJ only supports the first of these.

There is a proposal for EditorConfig to have a property for specifying the
number of imports required before using a wildcard import but it doesn’t seem
to have gained much traction (the proposed property doesn’t even have a name
yet). This is a shame, as it’s one of the few places where the style guide
differs from the IntelliJ defaults and it would be nice to not have to remind
every developer to change the setting manually.