Include nginx.conf in build files
Currently our nginx configuration is being ignored as it isn't part of the
files being pushed to CloudFoundry. Since we mostly just serve static files it
wasn't obvious until we're trying to make additional configuration changes.

This copies the nginx.conf into source files so middleman will copy it into the
build directory which is pushed to the PaaS.