Running Services as containers
This is intended as a conversation starter to discuss GDS guidance on
running containers.

Containers are being used to run lots of different types of services,
and there are often times when these services are not appropriate to run
on the PaaS, or the PaaS is currently not suitable for the provider
(ie GOV.UK Pay).

This commit serves to lay down some princibles for running your service
in this way.

It is explictly not related to development princibles on containers,
which should be a separate set of guidance.