Adding style guide for Node.js
After a [false start](https://github.com/alphagov/gds-way/pull/64) here is a second attempt
at a Node.js style guide. After a meeting with @maxcbc and @Nooshu we decided to
outline our preffered Formatting and Linting as using this will bring style inline
without requiring extensive documenation here.

We also outline how Node.js projects should be structured. We think a modular, feature
based structure is best because it emphasises the fact that code should be written in
a modular resuable way, we don't recommend an MVC style structure as this separation by
roles fragments the code and makes it more difficult to read and write.