adds new RSpec guide
This file is based on the state of repositories on alphagov at the moment, and guidance such as https://github.com/alphagov/publishing-e2e-tests/blob/improve-new-test-documentation/CONTRIBUTING.md#codingcommit-style.

As a new joiner, it's difficult to know the best 'expected' way of writing tests - particularly integration tests - as this information is missing from the styleguides.