Update 'git push -f' advice
The default for 'git pushes' changed with git 2.0 in 2014, so it only pushes the current branch, not to others too:
https://stackoverflow.com/questions/13148066/warning-push-default-is-unset-its-implicit-value-is-changing-in-git-2-0
So you don't have to mention the branch name in modern times.

It seems such a rare occurrence anyone will come across a version of git older in their normal dev workflow, and if they do, they will suffer issues with all their pushes, not just force pushes. So let's optimize guidance for the common case, reducing the cognitive cost of reading for everyone, and be bold :)

Co-authored-by: davidread <david.read@hackneyworkshop.com>
Co-authored-by: ChrisBAshton <christopher.ashton@digital.cabinet-office.gov.uk>