consolidate logging pages
we want to consolidate the pages about logs shipping, storage,
querying, and structuring to avoid fragmentation and allow future work
on explaining different logging methods.

This is initially a verbatim move of the content that was in the
"manuals" section into the main logging section within "how to operate a
service". We aim to further update this page to better reflect the
current state of logging systems at GDS in a future PR.