Document usage of standardx
GDS encountered a tricky problem with their standard choice when
standard 16 was released, this release pushed a new rule, no-var [1],
which issued warnings when projects used `var` over `let` or `const`.
This caused problems for a number of GDS projects where `var` is still
used as it required for JS to operate in IE < 11.

In order to resolve this problem the suggestion is to use standardx
which operates the same as standard but allows individual project rules.
This means that a project can continue using an up-to-date rule set for
standard while disabling any rules that are problematic (this is expected
to be `no-var` for most projects, but future problematic rules may
occur). This commit sets about documenting this decision and to
communicate the usage of standardx as a reasonable alternative to
standard.

When considering how to approach this problem we experimented with the
suggestion in this file of switching to a generic `.eslintrc` setup.
Doing this is actually rather complicated. We found that if we wanted to
share a rule set that built upon standard but with GDS customisations
then we had a rather arduous installation and configuration process (for
example the installation process went from `npm install --save-dev standard`
to `npm install --save-dev eslint eslint-config-gds eslint-plugin-node
eslint-plugin-import eslint-plugin-promise`) which seemed strange for a
one rule change, then there were a bunch of curiosities: should we disable
all rules that allow > ES5 functionality for consistency? why do browser
projects need to install various, seemingly unnecessary, plugins? should we
recommend one rule set for modern projects (standard) and another for
legacy (eslint-config-gds)?

Given these thoughts we decided that standardx was a simple pragmatic
choice that allowed us to move forward without being stuck on an old
version. It's quite conceivable that this problem could increase if
standard adopts further rules that have browser compatibility issues,
which may force us to reconsider if we need to share a config.
Hopefully, though, GDS will adopt the newer practices in JS and keep up
with the recommendations standard tries to enforce.

[1]: https://github.com/standard/standard/issues/633