We no longer have a DD Techops
The advice used to be that you had to check with the DD Techops before
using a new programming language.  I think part of the issue was a
concern that Techops would end up having to support/operate the
running service.

Anyway, Techops is no more, so maybe you should talk to your HoT
instead.