Make Koa the suggested framework for web apps
Koa was created by the team who made Express as a
better way to do the same things.

This changes our advice to suggest it over Express
because it's not a massive jump to work with it if
you're used to Express but:
- its use of async functions and promises instead
  of standard callbacks better matches our
  existing advice and makes errors simpler
- its use of async/await also makes it easier to
  change how control passes between middleware

It's also well documented, has a good ecosystem
and support and seems pretty trustworthy.