Remove the CDIO Pillars page
This page is now just information about the structure of CDIO, rather
than specific guidance relevant to the building or operating of
services.

The information on the Cyber team is likely useful, but is better
placed somewhere else. For now, any references to contacting the
Cyber team have been replaced with a link to their Slack channel.