Update programming languages guidance
Update the programming languages guidance.

The most significant change here is that DI is exploring Node.js more
as a backend language, because it is better suited to serverless systems
than Java is, due to the cold start times Java has.

(Note: it's possible to mitigate Java cold start times with GraalVM, but
we have so far not had success with this).