Fix broken links to the reliability engineering wiki
The Reliability Engineering wiki is being decommissioned. There are a
number of references to material in this wiki across the GDS Way pages
and these need to be fixed. In some cases there are also references to
the Reliability Engineering team, which no longer exists, so these also
need to be removed.

In some cases we can replace references to the reliability engineering
wiki with references to other parts of the GDS Way.

For the monitoring guidance, replace the reference to the Reliability
Engineering Prometheus service with a recommendation to use Prometheus
or CloudWatch Metrics for simpler use-cases.