Stop deploying the gds-tech-docs app
In 02870d3 the gds-way app was renamed from the old gds-tech-docs
app to give it a nicer URL. In its place the gds-tech-docs app was
deployed with an nginx configuration to redirect all requests to the new
gds-way URL.

The staticfile buildpack up to version 1.6.0 used ERB to generate its
nginx.conf, so it was possible to use ruby templating in nginx
configuration files, for example `listen <%= ENV["PORT"] %>;`.

However, in version 1.6.1 the staticfile buildpack stopped using ERB and
since then the gds-tech-docs app has failed to start with the error:

invalid log level "ENV["APP_ROOT"]" in /home/vcap/app/nginx/conf/nginx.conf:4

This could be fixed by switching to the nginx buildpack. However, the
redirector has been in place for almost 6 years and it is fair to say
it is no longer needed, so let's just stop deploying it instead.

https://github.com/cloudfoundry/staticfile-buildpack/blob/master/CHANGELOG