Remove HTML proofer config ignore_empty_alt
Ignore empty alt tag provides for:

> If true, ignores images with empty/missing alt tags (in other
words, <img alt> and <img alt=""> are valid; set this to false to
flag those)

Alt tags are a core bit of our accessibility requirements. We should be
enforcing their use. However even as I turn this off we have no failures
and so are already compliant. Changing this as a forward facing feature
to protect us against violating accessibility guidelines in future.