Make it clear that users in alphagov must have a GDS email address
the statement here was a bit vague. We are now tightening up the enforcement of
this policy, and so I'm amending the wording here.

I've slightly restructured the text as there was a mixture of 'advice about
configuring a repo' and 'how to set up your user account and gain access'. I've
created separate headings for those two topics