Fix broken link to ACL guidance
Despite the space between the link text and the `(ACL)` in brackets this is being interpreted as a markdown link with ‘ACL` as a destination, which does not exist. Rewording the link text to include the acronym and changing the name of the reference link fix this.

Also make it a relative URL, rather than including an outdated domain, and match the case of the target URL to avoid a redirect.