Update links to request an account tool
I have no idea why the markdown on this page is such a mess with so much duplication – I decided unpicking this mess is a job for another day.

I did however remove the link a non existent page on `gds-way-aws-preview.cloudapps.digital` (!?) because it wasn’t referenced anywhere and so wasn’t doing anything.