Java style guide: better guidance for IntelliJ ‘Optimize Imports’
Document the keyboard shortcut for IntelliJ IDEA’s ‘Optimize
Imports’ command and note it’s in the Code menu.

Provide better guidance on where one can select all modified Java
classes in IntelliJ IDEA and optimise the imports for just those.
Exactly where this can be done varies depending on whether or not
the non-modal commit interface is used and also differs slightly
depending on whether or not the new UI is used. Point out the
changes view in the commit tool window (non-modal) and the modified
files pane of the commit dialogue (modal), with links to the
IntelliJ IDEA docs.

This was discussed at the July 2024 Java community meeting.