Remove Go section under programing languages
If you are reading this commit by commit, no fear I add something back
in in the next commit.

Why remove this whole section at this level however?

Go's major reason to be at GDS was tied to PaaS, and PaaS has now
largely gone away.

I think the point the section makes about how there could be a use for
routing etc isn't a bad one, but I think that's far more likely to
happen today deep in the gubbins of AWS, in Infrastructure-as-code or in
languages like sucati for firewall rules.

So my vote here, is that we remove it wholesale. If there are future
ventures that come up with good uses for Go, they should be subject to
the bottom paragraph, raise it with your head of Tech and bring it back
formally with a good usecase.