Update threat-modelling.html.md.erb
Updated the link to Threat Modelling Scoring Template at the bottom of the page in response to Issue#979.
The new file was recreated from a working copy with all data and changes removed. This is now stored in the GOV.UK Technology Community Google Drive location and shared as a Viewer Only access with the link.