Java style guide: recommend upgrading Dropwizard to 3 then 4
Recommend upgrading from Dropwizard 2 to Dropwizard 3 then to
Dropwizard 4 rather than trying to go from 2 to 4 in one step.

Advise that some other dependencies that use Java EE will likely
have to be upgraded to Jakarta EE versions — but not all of them.

This was discussed at the January 2025 Java community meeting.