Remove Apple Silicon-specific setup instructions
I recently set up the GDS Way locally on a new GDS developer Macbook and
didn't have to do this. I can't see any recommendation to do this on the
Eventmachine github repo either, so I assume this issue is resolved now.

Let's remove it fo now, and if people have trouble setting it the
project up in future we can readd it.