Allow Dependabot to upgrade GitHub Actions for us
GitHub actions are just another way to stand on the shoulders of open
source code. Following the guidance in this very way we do SHA pinning.
That's great, it means we're guarded against supply chain attacks, but
it also means we don't get the latest and greatest code pulled every
time we run an Action.

When I was updating ruby recently I noticed some of the setup-ruby
actions were out of date and unaware of the most recent versions. In our
other static sites we use DepBot to monitor and suggest those like
dependencies. It's a bit of weekly noise, as these change a lot but in
my opinion is worth it for up-to-date actions, and to close off the risk
that we end up with very old pinned SHAs containing supply chain CVE
risks via guidance intended to mitigate the bleeding edge supply chain
risks of automated updates.

Dependabot making us aware seems part of the balanced approach to me