Turn on Dead Link checking as part of the test suite for this repo
Docs contain links, good docs contain links that work.
Dead link checking has been part of the test suite here before, looks
like in 2023 we turned it off temporarily and never spotted that we
needed to turn it back on.

This commit signals intend to bring it back in.

Immediately on doing this I'm greeted with over 100 errors. Though most
seem to be favicon related. I'll follow those up in subsequent commits