Update Links: New Semantic markup link in accessibility guidance
https://html.com/semantic-markup/ seems to work, but returns a 403
error, I suspect it's got some firewall rules that prevent certain user
agents or scrapers getting at it.

The content is fairly generic, so I've made a fairly unilateral decision
to replace it with an MDN link, that works and also is fairly good
guidance on the same topic.

Other guidance vendors may be available