Update Links: Update references to browsers we test in
Interesting one, so this changed to #browsers-to-test-in-from-september-2024
I suspect this anchor tag will break again, if it does I reckon we just
get rid of it and link to the page. It's a marginal scroll down to find
the right section.

But seeing as this is coming up to a year old, looks stable-ish.
Going to take a punt on just updating it.

If anyone is ever reading this thinking "oh why'd he do that", I'd
suggest dropping the anchor tag, I was on the fence when I did it.