Update Links: Add individual exceptions for 403-ing external Links
So plan here is to have two different sections.
I think the basic Ignore URLs should be RegEps that basically allow list
an entire site. In my view this is good for 403-ing sites where we know
there's an Auth requirement.

These should be accompanied with a 🔐 sign rendered next to the link so
folks know externally.

I've modified these to use %r{} syntax because I think it's clearer,
currently we had a mix of RegExp for GitHub and exact match strings.

I've also added another section for individual links. IMO these are a
1-to-1 match for a specific URL (and will need to be periodically
checked).

These are mostly sites that are indeed up, but 403 probably because they
have firewall rule that repel the scrape

I've tested the individual links they all work as of the time of this commit