Review page: Use GitHub - Clarify rules of access
This shouldn't be read as a change in terms, I'm clarifying rules a bit
by separating them into "what's common", and then details for our two
orgs.

Mostly this is a clarification to prevent confusion around
govuk-one-login, I think people here generally know the score for
`alphagov` already.

We will likely need to earmark this page for review in relation to DSIT
emails when we have a plan for changing over the access control bots