Review page: Use GitHub - Rearrange markdown links
This is a bit to taste but reviewing this was hard as the lines kept
running off the edge of my (fairly wide) screen.

So I've tried to condense the links into footnotes and in a couple of
places split the sentences, the result should be no visible change to
content, but easier to review and update links.