Improve docs for AWS MFA "You need permissions" error
SO many people miss the guidance about naming the MFA device correctly, and hit this error, and ask for help.

Steps are now numbered.
The guidance on how to fill the field correctly now has clearer warning section, more clearly associated with the step in question, not much earlier.
An image of the error message is shown on the step when it might appear, with guidance of how to deal with it.